requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
dash==3.3.0
dash-bootstrap-components==2.0.4
distro==1.9.0
fastapi==0.123.5
filelock==3.20.0
Flask==3.1.2
fsspec==2025.12.0
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
MarkupSafe==3.0.3
mpmath==1.3.0
narwhals==2.12.0
nest-asyncio==1.6.0
networkx==3.6
numpy==2.3.5
openai==2.9.0
packaging==25.0
plotly==6.5.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
retrying==1.4.2
safetensors==0.7.0
setuptools==80.9.0
sniffio==1.3.1
starlette==0.50.0
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.1+cpu
tqdm==4.67.1
transformers==4.57.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
Werkzeug==3.1.4
zipp==3.23.0
